+++ src/plaid/containers/dataset.py
@@ -931,11 +931,12 @@ class Dataset(object):
         for key, value in infos.items():
             self._infos[cat_key][key] = value
 
-    def set_infos(self, infos: dict[str, dict[str, str]]) -> None:
+    def set_infos(self, infos: dict[str, dict[str, str]], warn: bool = True) -> None:
         """Set information to the :class:`Dataset <plaid.containers.dataset.Dataset>`, overwriting the existing one.
 
         Args:
             infos (dict[str,dict[str,str]]): Information to associate with this data set (Dataset).
+            warn (bool, optional): If True, warns when replacing existing infos. Defaults to True.
 
         Raises:
             KeyError: Invalid category key format in provided infos.
@@ -963,7 +964,9 @@ class Dataset(object):
                             f"{info_key=} not among authorized keys. Maybe you want to try among these keys {AUTHORIZED_INFO_KEYS[cat_key]}"
                         )
 
-        if len(self._infos) > 0:
+        # Check if there are any non-plaid infos being replaced
+        has_user_infos = any(key != "plaid" for key in self._infos.keys())
+        if has_user_infos and warn:
             logger.warning("infos not empty, replacing it anyway")
         self._infos = copy.deepcopy(infos)